Set type of requested_frames of all mixer callbacks to int
The function pointer expects an int so using anything else is either misleading and an implicit cast at best or possibly undefined behavior at worst.
In edge cases, the callbacks using uint16_t can also overflow so this change prevents that bug as well.

+++ src/dos/cdrom.cpp
@@ -184,7 +184,7 @@ void CDROM_Interface_Physical::InitAudio()
 	thread = std::thread(&CDROM_Interface_Physical::CdReaderLoop, this);
 }
 
-void CDROM_Interface_Physical::CdAudioCallback(const uint16_t requested_frames)
+void CDROM_Interface_Physical::CdAudioCallback(const int requested_frames)
 {
 	assert(requested_frames > 0);
